move traits to own folder
